requirements.txt
https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-linux_armv7l.whl
pyopengl==3.1.5
picamera==1.13